# swagger/swagger.yaml
consumes:
    - application/json
definitions:
    ClusterConfiguration:
        properties:
            CreatedAt:
                format: date-time
                type: string
            DeletedAt:
                $ref: '#/definitions/DeletedAt'
            GroupName:
                type: string
            ID:
                format: uint64
                type: integer
            KubernetesConfiguration:
                items:
                    format: uint8
                    type: integer
                type: array
            UpdatedAt:
                format: date-time
                type: string
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/model
...
    CreateGroupRequest:
        properties:
            hostname:
                type: string
                x-go-name: Hostname
            name:
                type: string
                x-go-name: Name
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/group
...
    DeletedAt:
        $ref: '#/definitions/NullTime'
...
    Group:
        description: Group domain object defining a group
        properties:
            CreatedAt:
                format: date-time
                type: string
            DeletedAt:
                $ref: '#/definitions/DeletedAt'
            Hostname:
                type: string
            Name:
                type: string
            UpdatedAt:
                format: date-time
                type: string
            Users:
                items:
                    $ref: '#/definitions/User'
                type: array
            clusterConfiguration:
                $ref: '#/definitions/ClusterConfiguration'
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/model
...
    NullTime:
        description: |-
            NullTime implements the Scanner interface so
            it can be used as a scan destination, similar to NullString.
        properties:
            Time:
                format: date-time
                type: string
            Valid:
                type: boolean
        title: NullTime represents a time.Time that may be null.
        type: object
        x-go-package: database/sql
...
    User:
        description: User domain object defining a user
        properties:
            AdminGroups:
                items:
                    $ref: '#/definitions/Group'
                type: array
            CreatedAt:
                format: date-time
                type: string
            DeletedAt:
                $ref: '#/definitions/DeletedAt'
            Email:
                type: string
            Groups:
                items:
                    $ref: '#/definitions/Group'
                type: array
            ID:
                format: uint64
                type: integer
            UpdatedAt:
                format: date-time
                type: string
        type: object
        x-go-package: github.com/dhis2-sre/im-manager/pkg/model
...
info:
    contact:
        email: info@dhis2.org
        url: https://github.com/dhis2-sre/im-manager
    description: Instance Manager
    license:
        name: TODO
    title: Instance Manager
    version: 0.1.0
paths:
...
    /groups:
        post:
            description: Create a group...
            operationId: groupCreate
            parameters:
                - description: Refresh token request body parameter
                  in: body
                  name: Body
                  required: true
                  schema:
                    $ref: '#/definitions/CreateGroupRequest'
            responses:
                "201":
                    description: Group
                    schema:
                        $ref: '#/definitions/Group'
                "400":
                    $ref: '#/responses/Error'
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Create group
    /groups/{group}/cluster-configuration:
        post:
            description: |-
                Add a cluster configuration to a group. This will allow deploying to a remote cluster.
                Currently only configurations with embedded access tokens are support.
                The configuration needs to be encrypted using Mozilla Sops. Please see ./scripts/addClusterConfigToGroup.sh for an example of how this can be done.
            operationId: addClusterConfigurationToGroup
            parameters:
                - in: path
                  name: group
                  required: true
                  type: string
                  x-go-name: Group
                - description: SOPS encrypted Kubernetes configuration file
                  in: formData
                  name: Body
                  required: true
                  type: file
            responses:
                "201":
                    description: Group
                    schema:
                        $ref: '#/definitions/Group'
                "400":
                    $ref: '#/responses/Error'
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Add cluster configuration to group
    /groups/{group}/users/{userId}:
        post:
            description: Add a user to a group...
            operationId: addUserToGroup
            parameters:
                - in: path
                  name: group
                  required: true
                  type: string
                  x-go-name: Group
                - format: uint64
                  in: path
                  name: userId
                  required: true
                  type: integer
                  x-go-name: UserID
            responses:
                "201":
                    description: Group
                    schema:
                        $ref: '#/definitions/Group'
                "400":
                    $ref: '#/responses/Error'
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Add user to group
    /groups/{name}:
        get:
            description: Find a group by its name
            operationId: findGroupByName
            parameters:
                - in: path
                  name: name
                  required: true
                  type: string
                  x-go-name: Name
            responses:
                "200":
                    description: Group
                    schema:
                        $ref: '#/definitions/Group'
                "401":
                    $ref: '#/responses/Error'
                "403":
                    $ref: '#/responses/Error'
                "404":
                    $ref: '#/responses/Error'
                "415":
                    $ref: '#/responses/Error'
            security:
                - oauth2: []
            summary: Find group
...
produces:
    - application/json
responses:
...
    Error:
        description: ""
...
securityDefinitions:
    oauth2:
        flow: password
        tokenUrl: /tokens
        type: oauth2
swagger: "2.0"